+++ src/main/java/be/minelabs/entity/projectile/thrown/ChargedEntity.java
@@ -1,8 +1,12 @@
 package be.minelabs.entity.projectile.thrown;
 
+import be.minelabs.block.Blocks;
+import be.minelabs.block.blocks.TimeFreezeBlock;
 import be.minelabs.entity.BohrBlueprintEntity;
 import be.minelabs.entity.Entities;
 import be.minelabs.item.Items;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
 import net.minecraft.block.piston.PistonBehavior;
 import net.minecraft.entity.*;
 import net.minecraft.entity.damage.DamageSource;
@@ -15,16 +19,20 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.server.world.ServerWorld;
+import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.hit.EntityHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Box;
 import net.minecraft.util.math.Vec3d;
+import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.EmptyBlockView;
 import net.minecraft.world.World;
 
 import java.util.List;
 
 public abstract class ChargedEntity extends ThrownItemEntity {
     private static final TrackedData<Integer> CHARGE = DataTracker.registerData(ChargedEntity.class, TrackedDataHandlerRegistry.INTEGER);
+    private static final TrackedData<Boolean> STASIS_FIELD = DataTracker.registerData(ChargedEntity.class, TrackedDataHandlerRegistry.BOOLEAN);
 
     private Vec3d field = Vec3d.ZERO;
 
@@ -80,6 +88,12 @@ public abstract class ChargedEntity extends ThrownItemEntity {
     }
 
     @Override
+    public boolean canFreeze() {
+        // Don't want to freeze the entity in "powdered snow"
+        return false;
+    }
+
+    @Override
     public boolean canHit() {
         return true;
     }
@@ -112,6 +126,16 @@ public abstract class ChargedEntity extends ThrownItemEntity {
 
     @Override
     public void tick() {
+        Iterable<BlockPos> positions = BlockPos.iterateOutwards(getBlockPos(), e_radius, e_radius, e_radius);
+        for (BlockPos pos : positions) {
+            if (world.getBlockState(pos).isOf(Blocks.TIME_FREEZE_BLOCK)) {
+                if (world.getBlockState(pos).get(TimeFreezeBlock.LIT)) {
+                    //"Force" a stop
+                    setVelocity(Vec3d.ZERO);
+                }
+            }
+        }
+
         List<Entity> entities = world.getOtherEntities(this,
                 Box.of(this.getPos(), e_radius, e_radius, e_radius), entity -> entity instanceof ChargedEntity);
 
@@ -151,8 +175,21 @@ public abstract class ChargedEntity extends ThrownItemEntity {
     }
 
     @Override
+    protected void onBlockHit(BlockHitResult blockHitResult) {
+        BlockState state = world.getBlockState(blockHitResult.getBlockPos());
+        if (state.getCollisionShape(EmptyBlockView.INSTANCE, blockHitResult.getBlockPos(), ShapeContext.absent()) == VoxelShapes.empty()) {
+            // No collision keep moving
+            return;
+        }
+        Vec3d sideHit = Vec3d.of(blockHitResult.getSide().getVector()); // Side that the entity hit
+        addVelocity(getVelocity().multiply(sideHit));
+        super.onBlockHit(blockHitResult);
+    }
+
+    @Override
     protected void initDataTracker() {
         super.initDataTracker();
+        this.setNoGravity(true);
         dataTracker.startTracking(CHARGE, 0);
     }
 
@@ -160,12 +197,14 @@ public abstract class ChargedEntity extends ThrownItemEntity {
     public void readCustomDataFromNbt(NbtCompound nbt) {
         super.readCustomDataFromNbt(nbt);
         dataTracker.set(CHARGE, nbt.getInt("charge"));
+        dataTracker.set(STASIS_FIELD, nbt.getBoolean("stasis_frozen"));
     }
 
     @Override
     public void writeCustomDataToNbt(NbtCompound nbt) {
         super.writeCustomDataToNbt(nbt);
         nbt.putInt("charge", dataTracker.get(CHARGE));
+        nbt.putBoolean("stasis_frozen", dataTracker.get(STASIS_FIELD));
     }
 
     protected int getCharge() {
@@ -174,6 +213,14 @@ public abstract class ChargedEntity extends ThrownItemEntity {
 
     protected void setCharge(int charge) {
         dataTracker.set(CHARGE, charge);
+    }
+
+    public boolean isStasisFrozen() {
+        return dataTracker.get(STASIS_FIELD);
+    }
+
+    public void setStasisFrozen(boolean bool) {
+        dataTracker.set(STASIS_FIELD, bool);
     }
 
     public Vec3d getField() {
